refactor: Source code layout for CISKubeBenchReports
