+++ src/ciskubebenchreports/details.scss
